requirements.txt
openai>=1.12.0
pyyaml>=6.0
jsonschema>=4.20.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
anthropic>=0.50.0
google-genai>=1.0.0
